Refactor console output handling and improve error reporting
- Replace earlyprintln! macro with emergency_println! for better error handling in trap handlers.
- Update console_putchar to accept u8 instead of usize for consistency across architectures.
- Modify RISC-V boot and trap handlers to use println! instead of earlyprintln!.
- Introduce a BootConsoleBuffer to temporarily store console output during early boot before switching to runtime console.
- Enhance the Console trait to include a write_bytes method for raw byte output.
- Update various modules to utilize the new println! macro for logging and error messages.
- Remove deprecated print and println macros from stdio.rs to streamline output handling.
- Ensure that panic messages are printed using the new emergency_print method.

+++ os/src/arch/mock/lib.rs
@@ -10,7 +10,7 @@ pub fn restart() -> ! {
     }
 }
 
-pub fn console_putchar(_c: usize) {}
+pub fn console_putchar(_c: u8) {}
 
 pub fn console_getchar() -> usize {
     usize::MAX
